Fix H2A Sentinel Reader offset size

+++ src/LibSaber.Halo2A/IO/SentinelReader.cs
@@ -4,7 +4,7 @@ using LibSaber.Serialization;
 namespace LibSaber.Halo2A.IO
 {
 
-  public sealed class SentinelReader : SentinelReader<long>
+  public sealed class SentinelReader : SentinelReader<int>
   {
 
     public SentinelReader( NativeReader reader )
